# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Cadence": {
        "enum": [
          "DAILY",
          "WEEKLY",
          "MONTHLY",
          "QUARTERLY",
          "ANNUALLY"
        ],
        "title": "Cadence",
        "type": "string"
      },
      "CarePlan": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "item_name": {
            "title": "Item Name",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/RequestMetadata"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CarePlanStatus"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/CareTask"
            },
            "minItems": 1,
            "title": "Tasks",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "metadata",
          "plan_id",
          "session_id",
          "recommendation_id",
          "category",
          "item_name",
          "timezone",
          "status",
          "tasks",
          "created_at",
          "updated_at"
        ],
        "title": "CarePlan",
        "type": "object"
      },
      "CarePlanCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "preview_id": {
            "format": "uuid",
            "title": "Preview Id",
            "type": "string"
          }
        },
        "required": [
          "preview_id",
          "confirmed"
        ],
        "title": "CarePlanCreateRequest",
        "type": "object"
      },
      "CarePlanPreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "item_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Item Name",
            "type": "string"
          },
          "recommendation_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Recommendation Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "timezone": {
            "pattern": "^[A-Za-z_]+(?:/[A-Za-z0-9_+\\-]+)+$",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "recommendation_id",
          "category",
          "item_name",
          "start_date",
          "timezone"
        ],
        "title": "CarePlanPreviewRequest",
        "type": "object"
      },
      "CarePlanPreviewResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "confirmation_required": {
            "const": true,
            "default": true,
            "title": "Confirmation Required",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "item_name": {
            "title": "Item Name",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/RequestMetadata"
          },
          "preview_id": {
            "format": "uuid",
            "title": "Preview Id",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/CareTask"
            },
            "minItems": 1,
            "title": "Tasks",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "preview_id",
          "expires_at",
          "session_id",
          "recommendation_id",
          "category",
          "item_name",
          "timezone",
          "tasks"
        ],
        "title": "CarePlanPreviewResponse",
        "type": "object"
      },
      "CarePlanStatus": {
        "enum": [
          "ACTIVE",
          "PAUSED"
        ],
        "title": "CarePlanStatus",
        "type": "string"
      },
      "CarePlanUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CarePlanStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "CarePlanUpdateRequest",
        "type": "object"
      },
      "CareTask": {
        "additionalProperties": false,
        "properties": {
          "cadence": {
            "$ref": "#/components/schemas/Cadence"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "instructions": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Instructions",
            "type": "string"
          },
          "next_due_on": {
            "format": "date",
            "title": "Next Due On",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "instructions",
          "cadence",
          "next_due_on"
        ],
        "title": "CareTask",
        "type": "object"
      },
...
      "Category": {
        "enum": [
          "PLANT",
          "DOG",
          "CAT"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "contract_version": {
            "const": "v1",
            "default": "v1",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RequestMetadata": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionInfo"
          }
        },
        "required": [
          "request_id",
          "generated_at",
          "versions"
        ],
        "title": "RequestMetadata",
        "type": "object"
      },
...
      "VersionInfo": {
        "additionalProperties": false,
        "properties": {
          "api": {
            "default": "v1",
            "title": "Api",
            "type": "string"
          },
          "graph": {
            "default": "recommendation-graph-v1",
            "title": "Graph",
            "type": "string"
          },
          "knowledge": {
            "default": "fake-catalog-v1",
            "title": "Knowledge",
            "type": "string"
          },
          "model": {
            "default": "none",
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "default": "deterministic-explanations-v1",
            "title": "Prompt",
            "type": "string"
          },
          "rules": {
            "default": "rules-v1",
            "title": "Rules",
            "type": "string"
          },
          "scoring": {
            "default": "scoring-v1",
            "title": "Scoring",
            "type": "string"
          }
        },
        "title": "VersionInfo",
        "type": "object"
      }
    },
    "securitySchemes": {
      "GoogleIapAssertion": {
        "description": "Signed assertion injected by Google Cloud Identity-Aware Proxy.",
        "in": "header",
        "name": "X-Goog-IAP-JWT-Assertion",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned application API for deterministic Plant, Dog, and Cat recommendations and explicitly confirmed care plans.",
    "title": "Indoor Plant and Pet Advisor API",
    "version": "1.0.0",
    "x-api-version": "v1",
    "x-error-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/care-plans": {
      "post": {
        "operationId": "createCarePlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarePlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Persist an explicitly confirmed preview",
        "tags": [
          "care-plans"
        ]
      }
    },
    "/v1/care-plans/preview": {
      "post": {
        "operationId": "previewCarePlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarePlanPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlanPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Preview a care plan without saving it",
        "tags": [
          "care-plans"
        ]
      }
    },
    "/v1/care-plans/{plan_id}": {
      "get": {
        "operationId": "getCarePlan",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Retrieve a saved care plan",
        "tags": [
          "care-plans"
        ]
      },
      "patch": {
        "operationId": "updateCarePlan",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarePlanUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Pause or activate a care plan",
        "tags": [
          "care-plans"
        ]
      }
    },
    "/v1/care-plans/{plan_id}/tasks/{task_id}/complete": {
      "post": {
        "operationId": "completeCareTask",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or eligibility failure"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "GoogleIapAssertion": []
          }
        ],
        "summary": "Mark a care task complete",
        "tags": [
          "care-plans"
        ]
      }
    },
...
  },
...
}